Add files via upload
In this lab, I investigate strategies to identify pixels in image frames that have been impacted by cosmic rays through the construction of a median bias frame as well as the measurement of dark current for the DEIMOS detector.
Note: There are two Python notebooks that contain the code for my lab. The first notebook is titled “L3_AfrahAli” and the second notebook is titled “L3P3”. I separated my lab into two different notebooks because attempting to run all of the code from one single notebook was overworking my computer. Please review the Python notebooks in the listed order.